Added an events.js to export the RangeChangedEvent and VisibilityChangedEvent classes.
