+++ .changeset/mighty-suits-protect.md
@@ -1,0 +1,5 @@
+---
+'@lit-labs/virtualizer': patch
+---
+
+Added an events.js to enable exporting of RangeChangedEvent and VisibilityChangedEvent classes.
